requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.16.1
transformers>=4.32.1
accelerate>=0.22.0
peft>=0.4.0
pandas>=1.5.0
Pillow>=9.0.0